requirements.txt
biopython>=1.84
openai>=1.30.0
requests>=2.31.0

# Optional retrievers
rank-bm25>=0.2.2
sentence-transformers>=3.0.1
numpy>=1.26.0
torch>=2.1.0 